Refactor transforms to submodule
