+++ transform/coalesce_test.go
@@ -1,17 +1,13 @@
-package kazaam_test
+package transform
 
-import (
-	"testing"
-
-	"github.com/qntfy/kazaam"
-)
+import "testing"
 
 func TestCoalesce(t *testing.T) {
-	spec := `[{"operation": "coalesce", "spec": {"foo": ["rating.foo", "rating.primary"]}}]`
+	spec := `{"foo": ["rating.foo", "rating.primary"]}`
 	jsonOut := `{"foo":{"value":3},"rating":{"example":{"value":3},"primary":{"value":3}}}`
 
-	kazaamTransform, _ := kazaam.NewKazaam(spec)
-	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(testJSONInput)
+	cfg := getConfig(spec, "", false)
+	kazaamOut, _ := getTransformTestWrapper(Coalesce, cfg, testJSONInput)
 
 	if kazaamOut != jsonOut {
 		t.Error("Transformed data does not match expectation.")
@@ -22,10 +18,10 @@ func TestCoalesce(t *testing.T) {
 }
 
 func TestCoalesceWithRequire(t *testing.T) {
-	spec := `[{"operation": "coalesce", "spec": {"foo": ["rating.foo", "rating.primary"]},"require": true}]`
+	spec := `{"foo": ["rating.foo", "rating.primary"]}`
 
-	kazaamTransform, _ := kazaam.NewKazaam(spec)
-	_, err := kazaamTransform.TransformJSONStringToString(testJSONInput)
+	cfg := getConfig(spec, "", true)
+	_, err := getTransformTestWrapper(Coalesce, cfg, testJSONInput)
 
 	if err == nil {
 		t.Error("Coalesce does not support \"require\" and should throw an error.")
@@ -34,11 +30,11 @@ func TestCoalesceWithRequire(t *testing.T) {
 }
 
 func TestCoalesceWithMulti(t *testing.T) {
-	spec := `[{"operation": "coalesce", "spec": {"foo": ["rating.foo", "rating.primary"], "bar": ["rating.bar", "rating.example.value"]}}]`
+	spec := `{"foo": ["rating.foo", "rating.primary"], "bar": ["rating.bar", "rating.example.value"]}`
 	jsonOut := `{"bar":3,"foo":{"value":3},"rating":{"example":{"value":3},"primary":{"value":3}}}`
 
-	kazaamTransform, _ := kazaam.NewKazaam(spec)
-	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(testJSONInput)
+	cfg := getConfig(spec, "", false)
+	kazaamOut, _ := getTransformTestWrapper(Coalesce, cfg, testJSONInput)
 
 	if kazaamOut != jsonOut {
 		t.Error("Transformed data does not match expectation.")
@@ -49,11 +45,11 @@ func TestCoalesceWithMulti(t *testing.T) {
 }
 
 func TestCoalesceWithNotFound(t *testing.T) {
-	spec := `[{"operation": "coalesce", "spec": {"foo": ["rating.foo", "rating.bar", "ratings"]}}]`
+	spec := `{"foo": ["rating.foo", "rating.bar", "ratings"]}`
 	jsonOut := `{"rating":{"example":{"value":3},"primary":{"value":3}}}`
 
-	kazaamTransform, _ := kazaam.NewKazaam(spec)
-	kazaamOut, _ := kazaamTransform.TransformJSONStringToString(testJSONInput)
+	cfg := getConfig(spec, "", false)
+	kazaamOut, _ := getTransformTestWrapper(Coalesce, cfg, testJSONInput)
 
 	if kazaamOut != jsonOut {
 		t.Error("Transformed data does not match expectation.")
