+++ transform/extract.go
@@ -1,4 +1,4 @@
-package kazaam
+package transform
 
 import (
 	"fmt"
@@ -6,7 +6,8 @@ import (
 	simplejson "github.com/bitly/go-simplejson"
 )
 
-func transformExtract(spec *spec, data *simplejson.Json) (*simplejson.Json, error) {
+// Extract returns the specified path as the top-level object
+func Extract(spec *Config, data *simplejson.Json) (*simplejson.Json, error) {
 	outPath, ok := (*spec.Spec)["path"]
 	if !ok {
 		return nil, &Error{ErrMsg: fmt.Sprintf("Unable to get path"), ErrType: SpecError}
